+++ udata_front/theme/gouvfr/assets/js/composables/useOrganizationCertified.js
@@ -1,11 +1,17 @@
-import {ref, onMounted, unref} from 'vue';
+import {ref, onMounted, toValue} from 'vue';
 
+export const PUBLIC_SERVICE = "public-service";
+
+export const CERTIFIED = "certified";
+
+/**
+ * Know if an organization is certified or not
+ * @param {import("vue").MaybeRefOrGetter<import("../types").Organization>} organizationRef
+ */
 export default function useOrganizationCertified(organizationRef) {
-  const PUBLIC_SERVICE = "public-service";
-  const CERTIFIED = "certified";
   const organizationCertified = ref(false);
   const isOrganizationCertified = () => {
-    const organization = unref(organizationRef);
+    const organization = toValue(organizationRef);
     if(!organization) {
       organizationCertified.value = false;
       return;